backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
flask-pymongo>=2.3.0
pymongo[srv]>=4.8.0
requests>=2.32.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
numpy>=1.26.0
pandas>=2.1.0
scikit-learn>=1.4.0
nltk>=3.9.0
pythainlp>=5.0.0
python-dateutil>=2.9.0
dateparser>=1.2.0
sentence-transformers>=3.0.0
transformers>=4.40.0
torch>=2.3.0
python-dotenv>=1.0.0
